Add models and migrations for `tools` and `washes`
This commit adds the models and migrations for the tools and
washes apps. Washes can have many tools associated with them,
and each tool can be used in several washes.
Hence, a many-to-many relationship is implemented in the model
`Tools`, with `related_name = tools` so that it can be referred
back to.

Several other basic fields have been implemented for each model.